fix(svm): fix output typing for `lamports_to_sol` fn

+++ addons/svm/src/functions.rs
@@ -218,7 +218,7 @@ lazy_static! {
                 ],
                 output: {
                     documentation: "The number of lamports provided, represented as SOL.",
-                    typing: Type::integer()
+                    typing: Type::float()
                 },
             }
         }
